# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/plans/all": {
      "get": {
        "operationId": "AdminPlansController_getPlans",
        "summary": "Get all plans",
        "description": "Retrieve a paginated list of all plans with their prices and features",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in plan name and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/active": {
      "get": {
        "operationId": "AdminPlansController_getActivePlansWithPrices",
        "summary": "Get active plans with prices",
        "description": "Retrieve all active plans with their associated prices and features",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active plans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "prices": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      },
                      "features": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/{id}": {
      "get": {
        "operationId": "AdminPlansController_getPlanById",
        "summary": "Get plan by ID",
        "description": "Retrieve detailed information about a specific plan including prices and features",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Plan not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AdminPlansController_updatePlan",
        "summary": "Update an existing plan",
        "description": "Update plan properties such as name, description, or active status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Plan update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Plan not found"
          },
          "409": {
            "description": "Plan with this slug already exists for this location"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminPlansController_deletePlan",
        "summary": "Delete a plan",
        "description": "Soft delete a plan and all its associated features and prices",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan deleted successfully"
          },
          "400": {
            "description": "Failed to delete plan"
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/create": {
      "post": {
        "operationId": "AdminPlansController_createPlan",
        "summary": "Create a new plan",
        "description": "Create a new subscription plan with name, description and other properties",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Plan creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Plan with this slug already exists for this location"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/create-with-features": {
      "post": {
        "operationId": "AdminPlansController_createPlanWithFeatures",
        "summary": "Create a complete plan with features and prices",
        "description": "Create a new subscription plan with all features and pricing options in one request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Complete plan creation data with features and prices",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanWithFeaturesDto"
              },
              "examples": {
                "Premium Business Plan": {
                  "value": {
                    "officeLocationId": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Premium Business Plan",
                    "slug": "premium-business",
                    "description": "Comprehensive business plan with advanced features",
                    "imageUrl": "https://example.com/premium-plan.png",
                    "isActive": true,
                    "features": [
                      {
                        "featureId": "be2294ec-a909-462f-a81c-276d6ebbfb58",
                        "includedLimit": 10,
                        "unitPrice": 500,
                        "isActive": true
                      },
                      {
                        "featureId": "5e37ea6d-cba2-4413-8209-57b3b2a77035",
                        "includedLimit": null,
                        "unitPrice": null,
                        "isActive": true
                      }
                    ],
                    "prices": [
                      {
                        "billingCycle": "MONTHLY",
                        "amount": 4999,
                        "currency": "USD",
                        "description": "Monthly subscription"
                      },
                      {
                        "billingCycle": "YEARLY",
                        "amount": 49999,
                        "currency": "USD",
                        "description": "Annual subscription (save 17%)"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan with features and prices created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "officeLocation": {
                      "type": "object"
                    },
                    "features": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "featureId": {
                            "type": "string"
                          },
                          "includedLimit": {
                            "type": "number",
                            "nullable": true
                          },
                          "unitPrice": {
                            "type": "number",
                            "nullable": true
                          },
                          "feature": {
                            "type": "object"
                          }
                        }
                      }
                    },
                    "prices": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "billingCycle": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "currency": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or features not found"
          },
          "404": {
            "description": "Office location not found"
          },
          "409": {
            "description": "Plan with this slug already exists for this location"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/plan-prices/all": {
      "get": {
        "operationId": "AdminPlansController_getPlanPrices",
        "summary": "Get plan prices",
        "description": "Retrieve plan prices with optional filtering by plan, billing cycle, or currency",
        "parameters": [
          {
            "name": "currency",
            "required": false,
            "in": "query",
            "description": "Filter by currency code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "billingCycle",
            "required": false,
            "in": "query",
            "description": "Filter by billing cycle (monthly, yearly, etc.)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planId",
            "required": false,
            "in": "query",
            "description": "Filter by plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan prices retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "planId": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "number"
                      },
                      "currency": {
                        "type": "string"
                      },
                      "billingCycle": {
                        "type": "string"
                      },
                      "plan": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/plan-prices/{id}": {
      "get": {
        "operationId": "AdminPlansController_getPlanPriceById",
        "summary": "Get plan price by ID",
        "description": "Retrieve detailed information about a specific plan price",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan Price ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan price retrieved successfully"
          },
          "404": {
            "description": "Plan price not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AdminPlansController_updatePlanPrice",
        "summary": "Update plan price",
        "description": "Update plan price properties such as amount, currency, or billing cycle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan Price ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Plan price update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanPriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan price updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Plan price not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminPlansController_deletePlanPrice",
        "summary": "Delete plan price",
        "description": "Soft delete a plan price configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan Price ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan price deleted successfully"
          },
          "400": {
            "description": "Failed to delete plan price"
          },
          "404": {
            "description": "Plan price not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/plan-prices": {
      "post": {
        "operationId": "AdminPlansController_createPlanPrice",
        "summary": "Create plan price",
        "description": "Add a new price configuration for a plan with specific billing cycle and currency",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Plan price creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanPriceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan price created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Plan not found"
          },
          "409": {
            "description": "Price for this plan and billing cycle already exists"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/plans/plan-prices/{id}": {
      "get": {
        "operationId": "PlansController_getPlanPriceById",
        "summary": "Get plan price by ID",
        "description": "Retrieve detailed information about a specific plan price",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan Price ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan price retrieved successfully"
          },
          "404": {
            "description": "Plan price not found"
          }
        },
        "tags": [
          "Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Virtual Mail Hub",
    "description": "The Virtual Mail Hub API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePlanWithFeaturesDto": {
        "type": "object",
        "properties": {
          "plan": {
            "description": "Plan creation data",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            ]
          },
          "features": {
            "description": "Plan features to include",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePlanFeatureDto"
            }
          },
          "prices": {
            "description": "Plan prices to include",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePlanPriceDto"
            }
          }
        },
        "required": [
          "plan"
        ]
      },
...
    }
  }
}